pyproject.toml
[tool.poetry]
name = "amneziabot"
version = "0.0.1"
description = ""
authors = ["Amnezia VPN"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-split-settings = "^1.2.0"
python-decouple = "^3.6"
psycopg2-binary = "^2.9.5"
bcrypt = "^4.0.1"
ipython = "^8.6.0"
pytelegrambotapi = "^4.7.1"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
default_section = "FIRSTPARTY"